+++ client-app/src/@types/Activity.ts
