+++ client-app/src/services/ActivityService.ts
@@ -1,4 +1,4 @@
-import { Activity } from "../@types/ActivityType";
+import { Activity } from "../@types/Activity";
 import { request } from "./ApiService";
 
 const activityEndPoint = "/activities";
